requirements.txt
requests
python-dotenv
pyvis
networkx